move optimize_limits to data
